+++ config/core.config.properties
@@ -1,1 +1,3 @@
 enableConsole:true
+enableStemmer:true
+chooseLanguage:en
